requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
backcall==0.1.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.8.13
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
diff-match-patch==20121119
Django==2.1
django-celery==3.2.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.0.1
django-rest-swagger==2.1.2
django-reversion==3.0.0
djangorestframework==3.8.2
drf-generators==0.3.0
drf-openapi==1.3.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
kombu==3.0.37
Markdown==2.6.11
MarkupSafe==1.0
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.5
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycryptodomex==3.4.7
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
raven==6.9.0
requests==2.19.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
stream-python==2.10.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
Unipath==1.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0